Reorganise files (again...)
Get rid of the word "floating"
