+++ site/client/css/variables.scss
@@ -178,7 +178,7 @@ $controls-color: #3f9eff;
  */
 
 $zindex-input: 1;
-$zindex-floating-entity-select: 11;
+$zindex-global-entity-select: 11;
 $zindex-site-header: 12;
 $zindex-footnote: 15;
 $zindex-sidebar: 20;
