+++ site/client/owid.scss
@@ -36,7 +36,7 @@
 @import "css/pages/covid.scss";
 @import "css/pages/chart.scss";
 
-@import "css/components/floating-entity-select.scss";
+@import "css/components/global-entity-select.scss";
 @import "css/components/cookie-notice.scss";
 @import "css/components/site-tools.scss";
 @import "css/components/newsletter-subscription.scss";
