+++ site/server/formatting.tsx
@@ -531,14 +531,14 @@ export async function formatWordpressPost(
     const style =
         $("style").length === 1 ? `<style>${$("style").html()}</style>` : ""
 
-    // Render floating country selection component
+    // Render global country selection component
     $("*[data-entity-select]").each((_, el) => {
         const $el = $(el)
         const $section = $el.closest("section")
 
         $el.remove()
 
-        $("<div data-floating-entity-control />").insertAfter($section)
+        $("<div data-global-entity-control />").insertAfter($section)
     })
 
     return {
